voteValue input to 2-bigint
Following spec update from: https://github.com/vocdoni/docs/pull/70

voteValue is the hashed value of the *User* vote, composed by two big
integers:
- The raw user vote is a variable-length array of values and its values
do not need to be checked in the circuit. Furthermore, the values can be
encrypted.
- Since the encoded vote values may not fit into a constant number of
circuit inputs, we calculate a summary of the raw user vote using an
EVM-friendly hash function: `sha256(vote_bytes)`.  The output of the
sha256 hash is slightly larger than the field used in SNARKS, so we
split the hash output (32-bytes) into 2 16-byte arrays, take them as
integers (in little-endian), and use them as circuit inputs.