Update real nLevels used in circuit & js
Reason: circom uses nLevels to indicate what is nLevels-1. Until now
we were using for example 4 levels in circom & 3 in go code, to keep
both compatible, but this was 3 real levels, as the maximum leaf slots
were 2**3=8 leafs.  This commit updates the circom & js code to work
with nLevels+1, so we can use 4 nLevels in circom & js & go, and the
tree will have real 4 levels, allowing for 2**4=16 leafs.